requirements.txt
streamlit
neo4j
python-dotenv
google-generativeai